+++ lizard_riool/urls.py
@@ -4,7 +4,7 @@ from django.conf.urls.defaults import patterns
 from django.contrib import admin
 from django.contrib.auth.decorators import login_required
 from lizard_riool.views import Bar, FileView, PutFinder, PutList, \
-    SideProfileView, UploadView
+    SideProfileGraph, SideProfileView, UploadView, SideProfilePopup
 from lizard_ui.urls import debugmode_urlpatterns
 
 admin.autodiscover()
@@ -12,7 +12,9 @@ admin.autodiscover()
 urlpatterns = patterns('',
     (r'^beheer/$', login_required(FileView.as_view())),
     (r'^beheer/files/$', login_required(FileView.as_view(template_name="lizard_riool/files.html"))),
-    (r'^dwarsprofielen/$', login_required(SideProfileView.as_view())),
+    (r'^langsprofielen/$', login_required(SideProfileView.as_view())),
+    (r'^langsprofielen/graph/$', login_required(SideProfileGraph.as_view())),
+    (r'^langsprofielen/popup/$', login_required(SideProfilePopup.as_view())),
     (r'^upload/$', login_required(UploadView.as_view())),
     (r'^putten/$', login_required(PutList.as_view())),
     (r'^put/$', login_required(PutFinder.as_view())),